feat(aci): Switch metric resolution to radio
implements new designs

before

after
